requirements.txt
torch==2.1.2 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.16.2 --index-url https://download.pytorch.org/whl/cu121
numpy<2
typed-argument-parser
pytz
pandas
Pillow
tqdm
xformers==0.0.23.post1
flash-attn==2.7.2.post1
huggingface_hub==0.27.0
transformers
pyarrow
tensorboard
ftfy
regex
git+https://github.com/openai/CLIP.git
